completed reusable column component
