+++ ramp-up/src/Containers/Home/Table/Table.tsx
@@ -14,12 +14,12 @@ import {
 import { FixedColumns } from './TableColumns/FixedColumns/FixedColumns';
 import ErrorPopup from '../../../Components/ErrorNotification/ErrorNotification';
 import { useAppSelector, useAppDispatch } from '../../../Redux/hooks';
-import { emptyColumns, emptyRows } from './TableColumns/TableSkeletons/TableSkeletons';
+import { emptyColumns, emptyRows } from '../../../Components/TableSkeletons/TableSkeletons';
 import { Container, ButtonWrapper, StyledDataGrid, Title } from '../../../Utilities/TableStyles';
 import GridActionsColumn from './TableColumns/ActionColumn/ActionColumn';
 import { validateUser } from '../../../Utilities/ValidateUser';
 import { addUser, discardUser, saveUser } from '../../../Redux/user/userSlice';
-import { wait } from '@testing-library/user-event/dist/utils';
+import generateNewId from '../../../Utilities/generateRandomId';
 
 
 const Table = () => {
@@ -121,6 +121,15 @@ const Table = () => {
         });
     };
 
+    const handleAddClick = () => {
+        const id = generateNewId(rows)
+        setRows((oldRows) => [{ id, uid: id, name: '', gender: '', address: '', mobile: '', birthday: '', age: '', action: '', isNew: true }, ...oldRows,]);
+        dispatch(addUser({id:id, uid: id, name: '', gender: '', address: '', mobile: '', birthday: new Date(), age: Number(),isNew: true} ))
+        setRowModesModel((oldModel) => ({
+            ...oldModel, [id]: { mode: GridRowModes.Edit, fieldToFocus: 'uid' },
+        }));
+    };
+
     const columns: GridColDef[] = [
         ...FixedColumns, {
             field: 'actions',
@@ -141,16 +150,7 @@ const Table = () => {
         }
     ];
 
-    const maxId = rows.reduce((max, row) => (row.id > max ? row.id : max), 0);
-    const handleAddClick = () => {
-        const id = maxId + 1;
-        setRows((oldRows) => [{ id, uid: id, name: '', gender: '', address: '', mobile: '', birthday: '', age: '', action: '', isNew: true }, ...oldRows,]);
-        dispatch(addUser({id:id, uid: id, name: '', gender: '', address: '', mobile: '', birthday: new Date(), age: Number(),isNew: true} ))
-        setRowModesModel((oldModel) => ({
-            ...oldModel, [id]: { mode: GridRowModes.Edit, fieldToFocus: 'uid' },
-        }));
-
-    };
+  
 
     return (
         <Container>
